requirements.txt
functions-framework==3.3
pynacl==1.5.0
firebase_admin==6.2.0
pandas==1.5.3
pydantic==1.10.7
pyyaml==6.0
pylint==2.14.5
pylint-pydantic==0.1.8
google-cloud-tasks==2.13.1
protobuf==4.21.9
